feat(breadbox): Split out higher level functions into service layer (#169)
