+++ breadbox/breadbox/compute/analysis_tasks.py
@@ -27,8 +27,9 @@ from breadbox.schemas.custom_http_exception import ResourceNotFoundError, UserEr
 from breadbox.schemas.dataset import MatrixDatasetIn
 from breadbox.service import metadata as metadata_service
 
-from ..crud.types import get_dimension_type
+from ..crud.dimension_types import get_dimension_type
 from ..crud import dataset as dataset_crud
+from ..service import dataset as dataset_service
 from ..crud import group as group_crud
 from ..io import filestore_crud
 from .celery import app
@@ -482,7 +483,7 @@ def create_cell_line_group(
                 dataset_metadata=None,
                 dataset_md5=None,
             )
-            dataset_crud.add_matrix_dataset(
+            dataset_service.add_matrix_dataset(
                 db,
                 user,
                 dataset_in,
