+++ breadbox/breadbox/compute/dataset_uploads_tasks.py
@@ -19,10 +19,11 @@ from breadbox.schemas.custom_http_exception import ResourceNotFoundError
 from breadbox.models.dataset import DimensionType
 from fastapi import HTTPException
 from breadbox.io.filestore_crud import save_dataset_file
-from ..crud import dataset as dataset_crud
-from ..crud import types as type_crud
+from ..service import dataset as dataset_service
+from ..crud import dimension_types as type_crud
 from ..crud import group as group_crud
 from ..crud import data_type as data_type_crud
+from ..crud import dataset as dataset_crud
 from .dataset_tasks import db_context
 from ..api.uploads import construct_file_from_ids
 from ..io.data_validation import (
@@ -142,7 +143,7 @@ def dataset_upload(
             dataset_md5=dataset_params.dataset_md5,
         )
 
-        added_dataset = dataset_crud.add_matrix_dataset(
+        added_dataset = dataset_service.add_matrix_dataset(
             db,
             user,
             dataset_in,
@@ -187,7 +188,7 @@ def dataset_upload(
             dataset_metadata=dataset_params.dataset_metadata,
             dataset_md5=dataset_params.dataset_md5,
         )
-        added_dataset = dataset_crud.add_tabular_dataset(
+        added_dataset = dataset_service.add_tabular_dataset(
             db,
             user,
             dataset_in,
