+++ breadbox/breadbox/crud/dataset.py
@@ -1,6 +1,5 @@
 import logging
 from collections import defaultdict
-from dataclasses import dataclass
 from typing import Any, Dict, Optional, List, Type, Union, Tuple, Set
 from uuid import UUID, uuid4
 import warnings
@@ -12,8 +11,6 @@ from sqlalchemy.orm import aliased, with_polymorphic
 from breadbox.db.session import SessionWithUser
 from ..io.data_validation import dimension_label_df_schema
 from ..schemas.dataset import (
-    MatrixDatasetIn,
-    TabularDatasetIn,
     DimensionSearchIndexResponse,
     ColumnMetadata,
     UpdateDatasetParams,
@@ -26,7 +23,6 @@ from ..schemas.custom_http_exception import (
 )
 from breadbox.crud.access_control import user_has_access_to_group
 from breadbox.models.dataset import (
-    AnnotationType,
     Dataset,
     MatrixDataset,
     TabularDataset,
@@ -36,7 +32,6 @@ from breadbox.models.dataset import (
     DimensionSearchIndex,
     TabularColumn,
     TabularCell,
-    PropertyToIndex,
     ValueType,
     DimensionType,
     PrecomputedAssociation,
@@ -44,11 +39,8 @@ from breadbox.models.dataset import (
 from breadbox.crud.group import (
     get_group,
     get_groups_with_visible_contents,
-    TRANSIENT_GROUP_ID,
-    get_transient_group,
 )
 from breadbox.io.filestore_crud import delete_data_files
-from .metadata import cast_tabular_cell_value_type
 import typing
 
 log = logging.getLogger(__name__)
@@ -192,381 +184,7 @@ def get_dataset(
     return dataset
 
 
-def _get_dataset_group(
-    db: SessionWithUser, user: str, group_id: str, is_transient: bool
-):
-    # I don't really like this approach, but it's needed to support create_cell_line_group which
-    # creates some transient datasets. I think we haven't really thought through the role of groups
-    # and transient datasets. Every dataset needs a group, and so we have TRANSIENT_GROUP_ID that should be the
-    # owner of all transient datasets. Perhaps we should make group optional if isTransient is set? Or
-    # we could silently ignore the group_id in this case. For now, assert it's set appropriately
-
-    if is_transient:
-        if str(group_id) != TRANSIENT_GROUP_ID:
-            raise UserError(
-                f"If creating a transient dataset, the group must be set to {TRANSIENT_GROUP_ID}"
-            )
-        group = get_transient_group(db)
-    else:
-        group = get_group(db, user, group_id, write_access=True)
-    assert group is not None
-    return group
-
-
-def add_matrix_dataset(
-    db: SessionWithUser,
-    user: str,
-    dataset_in: MatrixDatasetIn,
-    feature_given_id_and_index_df: pd.DataFrame,
-    sample_given_id_and_index_df: pd.DataFrame,
-    feature_type: Optional[DimensionType],
-    sample_type: DimensionType,
-    short_name: Optional[str],
-    version: Optional[str],
-    description: Optional[str],
-):
-    group = _get_dataset_group(db, user, dataset_in.group_id, dataset_in.is_transient)
-
-    allowed_values = dataset_in.allowed_values
-
-    dataset = MatrixDataset(
-        id=dataset_in.id,
-        given_id=dataset_in.given_id,
-        name=dataset_in.name,
-        units=dataset_in.units,
-        feature_type_name=dataset_in.feature_type_name,
-        sample_type_name=dataset_in.sample_type_name,
-        data_type=dataset_in.data_type,
-        is_transient=dataset_in.is_transient,
-        group_id=group.id,
-        value_type=dataset_in.value_type,
-        priority=dataset_in.priority,
-        taiga_id=dataset_in.taiga_id,
-        allowed_values=allowed_values if allowed_values else None,
-        dataset_metadata=dataset_in.dataset_metadata,
-        md5_hash=dataset_in.dataset_md5,
-        short_name=short_name,
-        description=description,
-        version=version,
-    )
-    db.add(dataset)
-    db.flush()
-
-    _add_matrix_dataset_dimensions(
-        db=db,
-        index_and_given_id=feature_given_id_and_index_df,
-        dimension_subtype_cls=DatasetFeature,
-        dimension_type_name=feature_type.name if feature_type else None,
-        dataset=dataset,
-    )
-    _add_matrix_dataset_dimensions(
-        db=db,
-        index_and_given_id=sample_given_id_and_index_df,
-        dimension_subtype_cls=DatasetSample,
-        dimension_type_name=sample_type.name,
-        dataset=dataset,
-    )
-
-    db.flush()
-
-    return dataset
-
-
-@dataclass
-class PropertyMetadata:
-    value: str
-    dimension_id: str
-
-
-#    annotation_type: AnnotationType
-
-
-@dataclass
-class DatasetDimensionMetadata:
-    dimension_given_id: str
-    axis: str
-    type_name: str
-    label: str
-    property_metadata: Dict[str, PropertyMetadata]
-
-
-@dataclass
-class PropertyValuePair:
-    property: str
-    value: str
-
-
-# TODO: replace
-def get_metadata_by_dataset(
-    db: SessionWithUser, dataset: Dataset, properties_to_index: List[str]
-) -> pd.DataFrame:
-    from ..crud import types as types_crud
-
-    by_property = {}
-
-    index_type_name = dataset.index_type_name
-    for property in properties_to_index:
-        value_per_given_id = types_crud.get_dimension_type_metadata_col(
-            db, dimension_type_name=index_type_name, col_name=property
-        )
-
-        by_property[property] = pd.Series(value_per_given_id)
-
-    return pd.DataFrame(by_property)
-
-
-@dataclass
-class MetadataCacheEntry:
-    properties_to_index_df: pd.DataFrame
-    columns_metadata: Dict[str, ColumnMetadata]
-    label_by_given_id: Dict[str, str]
-
-
-class MetadataCache:
-    db: SessionWithUser
-    cache: Dict[str, MetadataCacheEntry]
-
-    def __init__(self, db: SessionWithUser):
-        self.cache = {}
-        self.db = db
-
-    def get(self, dimension_type_name: str):
-        from .types import get_given_id_to_dimension_id_mapping
-        from ..crud import types as types_crud
-
-        if dimension_type_name in self.cache:
-            entry = self.cache[dimension_type_name]
-        else:
-            dimension_type = types_crud.get_dimension_type(self.db, dimension_type_name)
-            assert dimension_type is not None
-            if dimension_type.dataset is None:
-                columns_metadata = {}
-                properties_to_index_df = pd.DataFrame({})
-                dimension_id_by_given_id = {}
-                label_by_given_id = {}
-            else:
-                print(dimension_type.dataset.index_type_name)
-
-                properties_to_index = [
-                    x.property for x in dimension_type.properties_to_index
-                ]
-
-                properties_to_index_df = get_metadata_by_dataset(
-                    db=self.db,
-                    dataset=dimension_type.dataset,
-                    properties_to_index=properties_to_index,
-                )
-
-                columns_metadata = dimension_type.dataset.columns_metadata
-
-                label_by_given_id = types_crud.get_dimension_type_metadata_col(
-                    self.db, dimension_type_name=dimension_type.name, col_name="label"
-                )
-
-            entry = MetadataCacheEntry(
-                properties_to_index_df=properties_to_index_df,
-                columns_metadata=columns_metadata,
-                label_by_given_id=label_by_given_id,
-            )
-            self.cache[dimension_type_name] = entry
-
-        return entry
-
-
-def get_property_value_pairs_for_given_id(
-    db, dimension_type_name: str, given_id: str, metadata_cache: MetadataCache
-) -> List[PropertyValuePair]:
-    dimension_search_index_rows = []
-
-    md_entry = metadata_cache.get(dimension_type_name)
-    properties_to_index_df = md_entry.properties_to_index_df
-    columns_metadata = md_entry.columns_metadata
-
-    # TODO: Rewrite
-    print("warning: inefficient. Rewrite")
-    rows_by_index = {}
-    for record in properties_to_index_df.to_records():
-        rows_by_index[record.index] = record
-
-    row = rows_by_index[given_id]
-
-    # if this ID is not specified, then just move on.
-    if row is not None:
-        # create one or more search index record for each column in properties_to_index_df
-        for property in properties_to_index_df.columns:
-            property_value = row[property]
-            cm = columns_metadata[property]
-
-            # now, check to see if we have a single value or multiple
-            if property_value is None:
-                values = []
-            elif cm.col_type == AnnotationType.list_strings:
-                values = cast_tabular_cell_value_type(
-                    property_value, AnnotationType.list_strings
-                )
-                assert isinstance(values, list)
-            else:
-                values = [property_value]
-
-            # are these values references to a keys in a different table
-            if cm.references is not None:
-                for value in values:
-                    for search_index_record in get_property_value_pairs_for_given_id(
-                        db,
-                        dimension_type_name=cm.references,
-                        given_id=value,
-                        metadata_cache=metadata_cache,
-                    ):
-                        # prefix the property name before adding to the index with the name
-                        # of the relationship that we traversed to get it. (this let's us
-                        # distinguish those properties which have the same name. For example, for
-                        # a compound, we'll get "alias" which is an alternative name for the
-                        # compound and "target.alias" which is an alternative name for the gene targeted
-                        # by the compound)
-                        dimension_search_index_rows.append(
-                            PropertyValuePair(
-                                property=f"{property}.{search_index_record.property}",
-                                value=search_index_record.value,
-                            )
-                        )
-            else:
-                # if not, then just add a search index record for each value
-                for value in values:
-                    dimension_search_index_rows.append(
-                        PropertyValuePair(property=property, value=value,),
-                    )
-
-    return dimension_search_index_rows
-
-
-def get_properties_to_index_query_for_metadata_dataset(
-    db: SessionWithUser, metadata_dataset_id: str, properties_to_index: List[str],
-):
-    filter_clauses = [
-        DimensionType.dataset_id == metadata_dataset_id,
-        Dimension.given_id.in_(properties_to_index),
-    ]
-
-    feature_query = (
-        db.query(TabularCell)
-        .join(TabularColumn, TabularCell.tabular_column_id == TabularColumn.id,)
-        .join(DimensionType, TabularColumn.dataset_id == DimensionType.dataset_id,)
-        .filter(and_(True, *filter_clauses))
-        .with_entities(
-            TabularCell.dimension_given_id,
-            TabularCell.value,
-            TabularCell.tabular_column_id,
-            TabularColumn.annotation_type,
-            DimensionType.id_column,
-            DimensionType.axis,
-            Dimension.dataset_dimension_type,
-            Dimension.given_id,
-            Dimension.id.label("dimension_id"),
-        )
-    )
-
-    return feature_query
-
-
-def _get_datatypes_referencing(db, dimension_type_name):
-    "Find the transitive closure over dimension types that reference other dimension types."
-    seen_names = set()
-
-    def _add_referencing_dimension_type(dimension_type_name: str):
-        assert isinstance(dimension_type_name, str)
-
-        seen_names.add(dimension_type_name)
-
-        # find any columns which reference this type
-        datasets_result = (
-            db.query(DimensionType)
-            .join(TabularDataset, TabularDataset.id == DimensionType.dataset_id)
-            .join(TabularColumn)
-            .filter(TabularColumn.references_dimension_type_name == dimension_type_name)
-            .with_entities(DimensionType.name,)
-            .all()
-        )
-
-        for (referencing_dimension_type_name,) in datasets_result:
-            if referencing_dimension_type_name not in seen_names:
-                # if we haven't seen those yet, add any columns that reference those too
-                _add_referencing_dimension_type(referencing_dimension_type_name)
-
-    _add_referencing_dimension_type(dimension_type_name)
-    return seen_names
-
-
-def _refresh_search_index_for_dimension_type(
-    db: SessionWithUser, dimension_type_name: str, metadata_cache: MetadataCache
-):
-    log.info(
-        "_refresh_search_index_for_dimension_type %s starting", dimension_type_name
-    )
-    from .types import get_dimension_type
-
-    dimension_type = get_dimension_type(db, dimension_type_name)
-    assert dimension_type is not None
-
-    if dimension_type.dataset_id is not None:
-        #        breakpoint()
-        assert user_has_access_to_group(
-            dimension_type.dataset.group, db.user, write_access=True
-        ), "Sign of user not having access to dataset needing to be indexed"
-
-        # before creating records in the search index for this dimension_type, clear out any previously existing records
-        log.info("_delete_search_index_records start")
-        _delete_search_index_records(db, dimension_type)
-        log.info("_delete_search_index_records complete")
-
-    dimension_search_index_rows = []
-
-    cache_entry = metadata_cache.get(dimension_type.name)
-
-    for given_id in cache_entry.properties_to_index_df.index:
-        for record in get_property_value_pairs_for_given_id(
-            db=db,
-            dimension_type_name=dimension_type.name,
-            given_id=given_id,
-            metadata_cache=metadata_cache,
-        ):
-            # if given_id in cache_entry.dimension_id_by_given_id:
-            dimension_search_index_rows.append(
-                DimensionSearchIndex(
-                    # dimension_id=cache_entry.dimension_id_by_given_id[given_id],
-                    property=record.property,
-                    value=record.value,
-                    group_id=dimension_type.dataset.group_id,
-                    dimension_type_name=dimension_type.name,
-                    dimension_given_id=given_id,
-                    label=cache_entry.label_by_given_id[given_id],
-                )
-            )
-
-    log.info("_refresh_search_index_for_dimension_type generated records. Writing...")
-
-    db.bulk_save_objects(dimension_search_index_rows)
-    log.info("_refresh_search_index_for_dimension_type complete")
-
-
-def populate_search_index(db: SessionWithUser, dimension_type: DimensionType):
-    impacted_dimension_types = _get_datatypes_referencing(db, dimension_type.name)
-
-    md = MetadataCache(db)
-    for impacted_dimension_types in impacted_dimension_types:
-        _refresh_search_index_for_dimension_type(db, impacted_dimension_types, md)
-    db.flush()
-
-
-def _delete_search_index_records(db: SessionWithUser, dimension_type: DimensionType):
-    db.query(DimensionSearchIndex).filter(
-        DimensionSearchIndex.dimension_type_name == dimension_type.name
-    ).delete()
-
-    db.flush()
-    return True
-
-
-def _add_matrix_dataset_dimensions(
+def add_matrix_dataset_dimensions(
     db: SessionWithUser,
     index_and_given_id: pd.DataFrame,
     dimension_subtype_cls: Union[Type[DatasetFeature], Type[DatasetSample]],
@@ -598,53 +216,7 @@ def _add_matrix_dataset_dimensions(
     db.flush()
 
 
-def add_tabular_dataset(
-    db: SessionWithUser,
-    user: str,
-    dataset_in: TabularDatasetIn,
-    data_df: pd.DataFrame,
-    columns_metadata: Dict[str, ColumnMetadata],
-    dimension_type: DimensionType,
-    short_name: Optional[str],
-    version: Optional[str],
-    description: Optional[str],
-):
-    # verify the id_column is present in the data frame before proceeding and is of type string
-    if dimension_type.id_column not in data_df.columns:
-        raise ValueError(
-            f'The dimension type "{dimension_type.name}" uses "{dimension_type.id_column}" for the ID, however that column is not present in the table. The actual columns were: {data_df.columns.to_list()}'
-        )
-
-    group = _get_dataset_group(db, user, dataset_in.group_id, dataset_in.is_transient)
-    dataset = TabularDataset(
-        id=dataset_in.id,
-        given_id=dataset_in.given_id,
-        name=dataset_in.name,
-        index_type_name=dataset_in.index_type_name,
-        data_type=dataset_in.data_type,
-        is_transient=dataset_in.is_transient,
-        group_id=group.id,
-        priority=dataset_in.priority,
-        taiga_id=dataset_in.taiga_id,
-        dataset_metadata=dataset_in.dataset_metadata,
-        md5_hash=dataset_in.dataset_md5,
-        short_name=short_name,
-        version=version,
-        description=description,
-    )
-    db.add(dataset)
-    db.flush()
-
-    _add_tabular_dimensions(
-        db, data_df, columns_metadata, dataset.id, dimension_type, group_id=group.id
-    )
-
-    db.flush()
-
-    return dataset
-
-
-def _add_tabular_dimensions(
+def add_tabular_dimensions(
     db: SessionWithUser,
     data_df: pd.DataFrame,
     columns_metadata: Dict[str, ColumnMetadata],
@@ -652,36 +224,11 @@ def _add_tabular_dimensions(
     dimension_type: DimensionType,
     group_id: str,
 ):
-    from .types import get_dimension_type
-
     """
     Adds tabular dataset dimensions to database.
     """
     dimensions = []
     values = []
-
-    assert (
-        dimension_type.id_column in data_df.columns
-    ), f"id column was specified as {dimension_type.id_column} but dataframe only had columns {data_df.columns}"
-
-    missing_metadata = set(data_df.columns).difference(columns_metadata)
-    if len(missing_metadata) > 0:
-        raise ValueError(
-            f"The following columns are missing metadata: {', '.join(missing_metadata)}"
-        )
-    extra_metadata = set(columns_metadata).difference(data_df.columns)
-    if len(extra_metadata) > 0:
-        raise ValueError(
-            f"The following columns had metadata but are not present in the table: {', '.join(extra_metadata)}"
-        )
-
-    for column_name, column_metadata in columns_metadata.items():
-        if column_metadata.references is not None:
-            dim_type = get_dimension_type(db, column_metadata.references)
-            if dim_type is None:
-                raise ValueError(
-                    f"The column {column_name} references {column_metadata.references} which does not exit"
-                )
 
     for col in data_df.columns:
         annotation_id = str(uuid4())
@@ -817,8 +364,6 @@ def get_dataset_dimension_search_index_entries(
         substrings,
         lambda table, substring: table.value.contains(substring, autoescape=True),
     )
-
-    from sqlalchemy import text
 
     search_index_query = (
         db.query(outer)
